comentarios

+++ src/main/java/br/com/casamagalhaes/workshop/desafio/model/Pedido.java
@@ -34,7 +34,7 @@ public class Pedido {
 
     private Double valorTotalProdutos;
 
-    private Long taxa;
+    private Double taxa;
 
     private Double valorTotal;
